Add comment on modulo bias in URL shortener
A comment was added to the `GenerateShortCode` method to explain the behavior of the modulo operation used. It highlights that the operation introduces a minor bias due to 256 not being evenly divisible by 62, slightly favoring the first 8 characters in the Base62 character set. The comment notes that this bias is negligible for URL shortening purposes and is outweighed by the simplicity of the implementation.